Rename jstree module
