+++ plugins/treeview/web_client/index.js
@@ -2,10 +2,12 @@ import * as utils from './utils';
 import * as types from './types';
 import * as children from './children';
 import * as views from './views';
+import jstree from './jstree';
 
 export {
     views,
     utils,
     types,
-    children
+    children,
+    jstree
 };
